Remove LOADER_FDT_SUPPORT as a Makefile variable.
LOADER_FDT_SUPPORT was used inconsistently in the tree. In some
places, it was used to control whether or not the user wanted FDT
included, and in other places it was a command to include
support. Remove it entirely. The former is now enabled -DWITH_FDT,
while the latter is controlled by Makefiles defining HAVE_FDT.

Supported by: Netflix


git-svn-id: https://svn.freebsd.org/base/head@325689 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f