Remove LOADER_ZFS_SUPPORT as a Makefile variable
LOADER_ZFS_SUPPORT is entirely used to select whether or not to
support ZFS in the loader. But it's not a user-servicable part (MK_ZFS
is what's used for that) Change it to the more conventional HAVE_ZFS
and move the ZFS support code into loader.mk. In addition, only build
ZFS libraries and boot loaders when ZFS is enabled.

Sponsored by: Netflix


git-svn-id: https://svn.freebsd.org/base/head@325690 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f