MFC r325441:
Extend the synopsis section of md(4) to look more like other manpages
of this kind.  Describe how to compile the driver into the kernel
and how to load it as a module.
This is useful for people using the MINIMAL kernel configuration file.

PR:			218610
Submitted by:		Harald Schmalzbauer (bugzilla.freebsd@omnilan.de)
Reviewed by:		noone (1 month inactivity)
MFC after:		2 weeks
Differential Revision:	https://reviews.freebsd.org/D12271


git-svn-id: https://svn.freebsd.org/base/stable/11@325996 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f