MFC r325813 (bz): Unbreak IPv6.
No longer return ENXIO when trying to send an IPv6 packet in
nicvf_sq_add_hdr_subdesc().
Restructure the code so that the upper layer protocol parts are
agnostic of the L3 protocol (and no longer specific to IPv4).
With this basic IPv6 packets go through.  We are still seeing
weird behaviour which needs further diagnosis.

PR:		223669


git-svn-id: https://svn.freebsd.org/base/stable/11@326062 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f