MFC r326137:
Ensure we check the program state set in the trap frame on arm and arm64.
This value may be set by userspace so we need to check it before using it.
If this is not done correctly on exception return the kernel may continue
in kernel mode with all registers set to a userspace controlled value. Fix
this by moving the check into set_mcontext, and also add the missing
sanitisation from the arm64 set_regs.

Discussed with:	security-officer@
Sponsored by:	DARPA, AFRL


git-svn-id: https://svn.freebsd.org/base/stable/11@326313 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f