MFC r325897: Improve the library dependencies helper script in src/tools.
Implement double pass of the relevant Makefiles. First make a list of
library names and directories and then scan for all the dependencies.
Spaces in directories in the source tree are not supported.

This avoids using hardcoded mappings between the library name
and the directory containing the library Makefile.

Add support for scanning contrib/ofed .

Bail out on any errors.

Sponsored by:	Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@326520 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f