Address a possible lost wakeup for gmirror events.
g_mirror_event_send() acquires the I/O queue lock to deliver a wakeup
to the worker thread, and this is done after enqueuing the event.
So it's sufficient to check the event queue before atomically releasing
the queue lock and going to sleep.

MFC after:	1 week
Sponsored by:	Dell EMC Isilon


git-svn-id: https://svn.freebsd.org/base/head@326798 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f